fix: Vagrant/Docker environment used wrong key for provisioning
Provisioning the docker environment failed for all playbooks using
gandalf instead of the default ansible user.

The underlying reason was that Vagrant was not allowed to use
the public key for gandalf from the ssh agent.

However, this is the intended design. The vagrant user is only
used to create the ansible user gandalf, which is then used
consistently across all providers.